Modularize APCpp
