+++ .gitmodules
@@ -1,3 +1,3 @@
-[submodule "lib/IXWebSocket"]
-	path = lib/IXWebSocket
-	url = https://github.com/machinezone/IXWebSocket
+[submodule "lib/APCpp"]
+	path = lib/APCpp
+	url = https://github.com/N00byKing/APCpp
